Make ds delete modal like the others

+++ frontend/src/composables/useDatastreams.ts
@@ -1,13 +1,15 @@
 import { useDatastreamStore } from '@/store/datastreams'
-import { onMounted, computed, ref, Ref } from 'vue'
+import { onMounted, computed, ref, Ref, watch } from 'vue'
 import { useThing } from './useThing'
 import { Datastream } from '@/types'
+import Notification from '@/store/notifications'
 
 export function useDatastreams(thingId: string) {
   const { is_owner } = useThing(thingId)
   const datastreamStore = useDatastreamStore()
   const selectedDatastream: Ref<Datastream | null> = ref(null)
   const isDeleteModalOpen = ref(false)
+  const deleteDatastreamInput = ref('')
 
   const visibleDatastreams = computed(() => {
     if (!datastreamStore.datastreams[thingId]) return []
@@ -26,7 +28,26 @@ export function useDatastreams(thingId: string) {
     isDeleteModalOpen.value = true
   }
 
+  function closeDeleteModal() {
+    selectedDatastream.value = null
+    // isDeleteModalOpen.value = false
+    deleteDatastreamInput.value = ''
+  }
+
+  watch(isDeleteModalOpen, (newValue) => {
+    if (newValue === false) {
+      closeDeleteModal()
+    }
+  })
+
   async function deleteDatastream() {
+    if (deleteDatastreamInput.value !== 'Delete') {
+      Notification.toast({
+        message: 'inputs do not match',
+        type: 'error',
+      })
+      return
+    }
     isDeleteModalOpen.value = false
     if (selectedDatastream.value) {
       await datastreamStore.deleteDatastream(
@@ -34,6 +55,7 @@ export function useDatastreams(thingId: string) {
         thingId
       )
     }
+    deleteDatastreamInput.value = ''
   }
 
   onMounted(async () => {
@@ -47,5 +69,6 @@ export function useDatastreams(thingId: string) {
     openDeleteModal,
     deleteDatastream,
     isDeleteModalOpen,
+    deleteDatastreamInput,
   }
 }
